add memcache based cache service

+++ src/main/java/com/github/chaijunkun/captcha/service/CacheService.java
@@ -8,6 +8,6 @@ public interface CacheService {
 
     boolean hasKey(String key);
 
-    void set(String key, String value, long expire);
+    void set(String key, String value, int expire);
 
 }
